Refactor EPICS classes into epics.py.
The cothread control system moves to cothread.py to allow optional
import.
