+++ pytac/lattice.py
@@ -258,49 +258,3 @@ class Lattice(object):
                                    "family")
         for element, value in zip(elements, values):
             element.set_value(field, value)
-
-
-class EpicsLattice(Lattice):
-
-    def __init__(self, name, energy, epics_cs):
-        """
-        control_system (ControlSystem): The control system used to store
-        the values on a PV.
-        """
-        super(EpicsLattice, self).__init__(name, energy)
-        self._cs = epics_cs
-
-    def get_pv_names(self, family, field, handle):
-        """Get all PV names for a specific family, field, and handle.
-
-        Assume that the elements are EpicsElements that have the get_pv_name()
-        method.
-
-        Args:
-            family (str): requested family.
-            field (str): requested field.
-            handle (str): pytac.RB or pytac.SP.
-
-        Returns:
-            list: list of PV names.
-        """
-        elements = self.get_elements(family)
-        pv_names = []
-        for element in elements:
-            pv_names.append(element.get_pv_name(field, handle))
-        return pv_names
-
-    def get_values(self, family, field, handle, dtype=None):
-        pv_names = self.get_pv_names(family, field, handle)
-        values = self._cs.get(pv_names)
-        if dtype is not None:
-            values = numpy.array(values, dtype=dtype)
-        return values
-
-    def set_values(self, family, field, values):
-        pv_names = self.get_pv_names(family, field, 'setpoint')
-        if len(pv_names) != len(values):
-            raise LatticeException("Number of elements in given array must be"
-                                   " equal to the number of elements in the "
-                                   "family")
-        self._cs.put(pv_names, values)
